split config.py
