+++ src/sts/config/__init__.py
@@ -1,0 +1,26 @@
+from sts.config.loader import get_app_settings
+from sts.config.buckets_map import BucketsMap, create_buckets_map, get_buckets_map
+from sts.config.models import (
+    AppSettings,
+    BucketSettings,
+    ImageSize,
+    S3HttpRetries,
+    S3HttpSettings,
+    S3Settings,
+)
+
+__all__ = [
+    # configuration models
+    "AppSettings",
+    "BucketSettings",
+    "ImageSize",
+    "S3HttpRetries",
+    "S3HttpSettings",
+    "S3Settings",
+    # methods
+    "get_app_settings",
+    "create_buckets_map",
+    "get_buckets_map",
+    # buckets model
+    "BucketsMap",
+]
